Always ref result of g_io_stream_get_{input,output}_stream.
Sometimes we wrap those in other streams, sometimes not. This makes prpls inconsistent with each other on whether they unref their `GInputStream`/`GOutputStream` in prpl data during shutdown. Thus, ref it when it's not wrapped, so we can _always_ do an unref in the cleanup.

This fixes a read-after-free Jabber shutdown, and presumably one in Novell, since it appears to do the same thing.

Testing Done:
Ran with valgrind and see no invalid reads in jabber shutdown. Novell was not tested, but seems similar.

Reviewed at https://reviews.imfreedom.org/r/558/