Only allow gplugin version 0.32.0.
This will make it easier for everyone to update subprojects and the like by running `meson subprojects update` from the source directory. We'll need to add this to the documentation I'm just not sure exactly where right now.

Testing Done:
Ran `meson subprojects update` from the source directory with `subprojects/gplugin` on the default branch and the v0.31.0 tag. Verified that meson will update to whatever revision is specified and if the revision doesn't exist it will pull from the remote.

Reviewed at https://reviews.imfreedom.org/r/812/