Don't error out when trying to remove a password that isn't stored.
Both wincred and secretservice return an error if you try to delete a password
that they don't know about. We don't care about this scenario and just want to
make sure the password is gone, so we changed the expected behavior to be as
such and updated wincred and secret service for the change.

I also fixed the NULL deference which is how I stumbled across the issue in the
first place.

Testing Done:
* Verfied the issue was fixed in secretservice.
* Verfied the issue didn't exist in kwallet.

Reviewed at https://reviews.imfreedom.org/r/901/