Add icon-name, icon-search-path, and icon-resource-path to PurpleProtocol
These are now used by pidgin_create_icon_from_protocol which will attempt to
load the icon for the protocol from the default GtkIconTheme if the protocol
has a non NULL value for icon-name.

pidgin_ui_init was also updated to add icon-search-path and icon-resource-path
to the default GtkIconTheme when a protocol is registered. It also scans all of
the protocols that have been already been registered as that happens before the
ui is initialized.

Testing Done:
Removed old local icon and then ran with an inverted icon to make sure pidgin was displaying the icon from the resource.

Reviewed at https://reviews.imfreedom.org/r/790/