Remove usage of pidgin_conv_is_hidden from transparency plugin
This plugin is only used on Windows so that's probably why we didn't notice
that it used this function when we deprecated it a few weeks ago.

Testing Done:
Compiled on Windows and verified the plugin can be enabled without crashing. Couldn't test it with a conversation since Pidgin3 on Windows crashes for me when I try to add a new account.

Reviewed at https://reviews.imfreedom.org/r/1139/