Remove a bunch of unused pixmaps
Testing Done:
Compiled and ran, didn't see anything missing.

It's worth noting that we've been using the logo from the resource for a long time, and the other icons were removed when we purge the `GtkStock` stuff, but there's existing items in the default icon theme for them anyways. Also we nuked the system tray stuff forever ago to.

Reviewed at https://reviews.imfreedom.org/r/1336/