Create a new PidginAccountManager that replaces the manually built one.
The old code was left to keep this change as clean as possible and will be
removed afterwards.

Drag and drop as been removed as it's completely different in GTK4 and I didn't
want to waste time on it since it's going to be deleted when it's rewritten.

I also didn't implement the double click to modify account as that's different
in GTK4 and we don't yet have new Event Controller for it in GTK3.

Support for selecting multiple accounts was dropped and now only one account may be selected at a time.

Finally, I changed the title of the "Delete" button to "Remove" as "Remove"
seems more clear.

Testing Done:
Used the previous dialog to add, remove, and modify accounts and verified they were updated appropriately.
Also started pidgin without any accounts enabled, to verify that the new dialog was shown instead of the old.
Tested the add, modify, remove, and close buttons.
Tested that the modify and remove buttons are only sensitive when an account is selected.

Bugs closed: PIDGIN-17589

Reviewed at https://reviews.imfreedom.org/r/1318/