Replace PidginAccountsDisabledMenu with PidginInactiveAccountsMenu which is a GMenuModel subclass.
This gives us a bit more control as we are always in control of the attributes,
so we don't need to modify the menu in place. This also made it trivial to add
the protocol icons which are very helpful when you have the same user name on
multiple protocols.

Testing Done:
Enabled and disabled two accounts a lot while under valgrind.

Reviewed at https://reviews.imfreedom.org/r/1366/