Start implementing the conversation window actions.
Still not sure how we want to handle add/remove and block/unblock, but this is
small set of things to get us started.

Testing Done:
Opened a conversation windown with multiple ims and a chat and verified that items sensitivity was set correctly depending on what was selected and made sure the actions work as well.

Side note, setting aliases on chats appears to have been broken outside of this change.

Reviewed at https://reviews.imfreedom.org/r/1563/