Remove the focus signal from the conversation pane.
I _think_ this was trying to make any focus goto the editor, but boxes don't
support this signal anymore and we have some more shuffling to do here so it's
not really worth the time to fix it at the moment.

Testing Done:
Ran and verified the `GWarning` was gone.

Reviewed at https://reviews.imfreedom.org/r/1666/