Rename PurpleContact to PurpleMetaContact.
This is to help facilitate the new blist api that's still being designed but we
wanted to use this name for an object. A later review request will rename
contact.[ch] which was avoided for now as it would make the diff harder to read.

Testing Done:
Ran and verified that contacts showed up.

Reviewed at https://reviews.imfreedom.org/r/1770/