Remove auto-replies
Auto-replies only work if the protocol supports them, and at the time of this
writing the only protocol that sets that flag is Zephyr which is barely used.

Testing Done:
Opened the preferences and verified the away page loaded fine.

Bugs closed: PIDGIN-17684

Reviewed at https://reviews.imfreedom.org/r/1867/