Remove the default formatting preference.
It turns out only a very few protocols still support the kind of formatting that
AIM/ICQ/Yahoo! used to support so there's not a whole lot of reasons to keep
this around.

Testing Done:
Opened preferences, made sure the conversations pane was fine and that there were no errors or warnings in the console.

Reviewed at https://reviews.imfreedom.org/r/1864/