Add the require password option to PidginAccountEditor
Testing Done:
Created an IRC account and set the property. Verified it was saved in `accounts.xml`. Reopened Pidgin, opened the account editor with the account connected, and verified that the switch was in the on position and insensitive. Disabled account and verified I could change the setting again.

Bugs closed: PIDGIN-17689

Reviewed at https://reviews.imfreedom.org/r/1877/