Use purple_debug_info on unknown irc messages
Previously, this was using purple_debug_error which maps to g_critical and makes
it annoying to run with G_DEBUG=fatal-warnings when connecting an IRC account.

Testing Done:
ran via `G_DEBUG=fatal-warnings gdb pidgin3` from a devenv and verified `pidgin3` didn't get paused when connecting to libera when it got `PRIVMSG`'s to `*` which was triggering the previous pause in `gdb`.

Reviewed at https://reviews.imfreedom.org/r/1895/