Convert PidginProxyPrefs to GSettings
This is just the preferences nothing has been done to the actual global proxy options. To configure those you'll need to edit prefs.xml manually until they are updated to gsettings.

I also made the rows in the preferences not focusable and split up `pidgin_proxy_prefs_init` for readability.

Testing Done:
Set all values via the preference page and verified they were saved properly in the prefs file. I also verified the mappings for the type and set all of them via the preference page and verified that it was saved properly in the preference file.

I also manually modified the preferences file and verified that the changes were made immediately in the running program. I also set an invalid value for the type enum, which then fell back to it's default value of "No Proxy".

Reviewed at https://reviews.imfreedom.org/r/1990/