Create PurpleUi that merges PurpleCoreUiOps and PurpleUiInfo
This created a lot more refactoring that I imagined, but it's mostly just moving stuff around.

I skipped the unit tests for now because we already have a `test_ui.[ch]` in the tests directory which we can/should move to a separate library or something, so it's available outside of the `libpurple/tests/` directory.

Testing Done:
Ran the unit tests, launched pidgin3, and launched finch3 with out issue.

Bugs closed: PIDGIN-17697

Reviewed at https://reviews.imfreedom.org/r/1925/