Port some of gg to libsoup3
This does not do the file transfer stuff as I'm not entirely sure of the flow yet, and also don't know how best to do libsoup3 for streaming uploads.

I did add a `cancellable` to the file transfer code though, to be used later.

Testing Done:
Compiled with soup2 and with soup3.

Reviewed at https://reviews.imfreedom.org/r/2011/