Fix SoupMessage leaks when using new libsoup3 API
The new API takes a ref, not ownership, of the sent message, so we need to unref our copy.

The compatibility layer for libsoup2 similarly also takes a ref, so this is also leaking on libsoup2 now.

Also fix an incorrect data parameter in the gnttinyurl plugin (all the rest of the code was already changed expecting `msg` as the callback data.)

Testing Done:
Compile only.

Reviewed at https://reviews.imfreedom.org/r/2017/