Fix string size specifier in debug prints
The `%*s` specifies the *width* of the string, but this was intended to specify
the maximum number of characters (in case the `GBytes` was unterminated), which is `%.*s`.

Testing Done:
Compiled only, though I did write a small test program with `printf` to confirm that this worked the intended way.

Reviewed at https://reviews.imfreedom.org/r/2023/