Port the rest of gg to libsoup3
However, I did not implement sending files, as libsoup requires a `GInputStream`, and that's some work to set up in a streaming manner. As we want to move the file transfer stuff to GIO, I think we'd wait to do that fully until then.

Testing Done:
Compiled with libsoup2 and with libsoup3.

Reviewed at https://reviews.imfreedom.org/r/2026/