Add purple_account_manager_get_connected and port libpurple to it
Generally speaking, purple_connections_get_all is used to get a list of accounts
that are currently connected and then act upon those accounts.

Instead of having a connections API for that, we can just put it into the
account manager which can query the accounts it knows about and check if they
are connected or not making the account the source of truth for whether it is
connected or not.

This also has the benefit of treating the connection as just a data object that
belongs to the account which feels a lot more natural than the existing API.

Testing Done:
Ran and let a bonjour account go idle and verified via another instance of pidgin.

Reviewed at https://reviews.imfreedom.org/r/2031/