Propagate the PurpleAccount::notify signal via PurpleAccountManager::account-changed
This new signal supports details and works just like notify does on the account
instances, but it will be emitted for any account that the manager instance
knows about which means we can remove the old purple signals.

Testing Done:
Enabled and disabled accounts in Pidgin via the menus to make sure things were working right, and verified via the console that the accounts were being enabled and disabled.
Ran Finch, but I ran into the infinite loop bug in the buddy list clean when disabling an account which is a known issue.

Reviewed at https://reviews.imfreedom.org/r/2032/