Remove gtkconn.[ch]
This was supposed to be doing the auto reconnect stuff, but apparently we've
been reconnecting on accident when we set an accounts idle status. This forces
the status of the account from offline to an online status, which then makes it
reconnect.

This is clearly a bug, but a happy accident for now as we're working on
re-architecting auto-reconnect.

Testing Done:
Use the demo account action to trigger a temporary connection failure and was surprised when the account reconnected.

Reviewed at https://reviews.imfreedom.org/r/2040/