Remove support for detecting Windows suspend/resume
This was used to disconnect and reconnect accounts before suspend and after
resume. This was never necessary on non-Windows so we suspect that the libc
interface that we had for Windows sockets may have been the real culprit here.

Testing Done:
Compiled and ran on Windows

Reviewed at https://reviews.imfreedom.org/r/2043/