Remove purple_account_request_add
This behavior has been replicated by purple_add_contact_request_new and the
notification API. The only consume of this function was the demo protocol
plugin, so this was updated to use the aforementioned API.

Testing Done:
Used the demo protocol plugin's account action to test it. Also fixed an off-by-one in that action.

Reviewed at https://reviews.imfreedom.org/r/2050/