Use a checksum of the protocol id and account username for accounts without ids
This still isn't bullet proof, but it's better than just a UUID4. Any change to
a username, or switch protocols will potentially orphan lots of records, but
this is a step in the right direction.

In the future I'm thinking account names might be better, but the jury is still
out on that one.

Testing Done:
Created a new account and verified that it got a sha256 digest as its id in `accounts.xml`.

Reviewed at https://reviews.imfreedom.org/r/2052/