Don't allow duplicate tags in PurpleTags
Originally I was thinking it might be useful to store counter objects in the
tags, but that adds a bunch of complexity to the serialization layer and tag
users can just get the existing tag, and then increment the value themselves.

Testing Done:
Ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2053/