Fix actions in PidginPluginsMenu
The issue was that G_MENU_ATTRIBUTE_ACTION_NAMESPACE only works if the menu
item is a section. So we create a new menu and add the plugins menu to it in
a section that has the property set and everything works.

Testing Done:
Used the actions from idle maker.

Bugs closed: PIDGIN-17708

Reviewed at https://reviews.imfreedom.org/r/2056/