Add a contact property to PurpleAccount
This contact can be used by the protocol to store additional information about
the account, including the user id, alias, etc. When the protocol connects, it
will be responsible for adding this contact to the ContactManager once it knows
that the id is correct.

This will also be used in the future when PurpleMessage gets migrated to using
PurpleContact's as well.

Testing Done:
Connected an IRCv3 and demo accounts.

Reviewed at https://reviews.imfreedom.org/r/2064/