Make PurpleContact derivable
We want to make PurpleContact derivable so that we can subclass it with
PurpleAccount. This will make things easier for everyone as we start moving
everything to PurpleContact as an account will just be a contact.

Testing Done:
Ran the unit tests and built the docs.

Reviewed at https://reviews.imfreedom.org/r/2081/