Remove purple_markup_escape_text
This was originally added because g_markup_escape_text turns ' in to &apos;
which isn't an element in HTML 4 and apparently it wasn't rendered right in
IE7. Those days are long gone, so there's no reason we need this anymore.

Testing Done:
Compiled and ran the unit tests

Reviewed at https://reviews.imfreedom.org/r/2096/