Add a setting-changed signal to PurpleAccount and propagate it from PurpleAccountManager
Also add a freeze and thaw mechanism to make sure state is consistent when the
signal is emitted.

Add a `GParamSpec` parameter to the `account-changed` signal as well, so that users can determine which property changed.

Testing Done:
Used `g_message` in the propagator to verify that everything was behaving.

Reviewed at https://reviews.imfreedom.org/r/2060/