Fix a compilation error on windows and many warnings
SIG_ACK is macro for handling signals on windows. We just want to use it as a
constant, so we #undef if we see it.

Mark a bunch of parameters as unused.

Fix a bunch of warnings about type casting function pointers.

Testing Done:
Compiled and ran on windows.

Reviewed at https://reviews.imfreedom.org/r/2106/