Make PurpleAccount subclass PurpleContactInfo
This allows accounts to be treated as PurpleContactInfo without duplication.

The existing api was left with deprecation flags so we can knock them out file by file.

Buddy Icons will replaced once /r/2092/ is merged.

Presence is a little goofy for now, because it still depends on PurpleAccountPresence which I want to remove in a simplification of the status api. I considered overriding the propert, but since `purple_contact_info_get_presence` returns its instance, this won't work.

Testing Done:
Ran the unit tests and connected the demo protocol and an XMPP account that had contacts in the list.

Reviewed at https://reviews.imfreedom.org/r/2113/