Remove the icon from PurpleIMConversation
The intent is to move this functionality to PurpleConversation once PurpleAvatar
has laned, so this just moves the exist code out of the way.

Testing Done:
Ran the tests and ran with demo and xmmp accounts. Avatars are busted in the info pane at the moment.

Reviewed at https://reviews.imfreedom.org/r/2114/