Add a name-for-display property to PurpleContactInfo
This is a readonly property that is updated when the other properties change.
I also tweaked priorities a bit, a contact info's alias is now the highest
priority to allow users to set an alias on a person, but then also override it
at the contact info level. The use case here is person has an alias of "Pidgy",
but Pidgy's slack account is aliased to "Pidgy (work)".

Testing Done:
Ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2151/