Add animation support to PurpleAvatar
This drastically changes the way PurpleAvatar works. Previously the idea was
that developers would create a PurpleAvatar and then add it to the
PurpleAvatarManager (which does not exist). But these changes are going to make
it so that developers will either lookup avatars or ask the manager to create
an avatar based on their data.

Testing Done:
Ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2141/