Create a new PidginContactList widget that shows the PurplePersons that PurpleContactManager is tracking
This has a few caveats with it.

First of all, we're using closures just to make stuff work for now since we can't bind properties. A side effect of that, is that the list won't update when the property changes on the Person/Contact. That will be fixed later as we clean up a bunch of stuff like porting PidginAvatar to PurpleAvatar.

For the text, I skipped the status messages and idle times for now because they won't update as their not bound properties and will require cleaning up the status API to make that work at all.

Person/Contact don't have a `name-for-display` property, so right now that's implemented as a closure as well.

I also skipped search and sorting for now, to keep the review size small.

Regardless, this looks pretty snazzy if I do say so!

Testing Done:
Ran the program and verified stuff was working.

Reviewed at https://reviews.imfreedom.org/r/2146/