Replace GtkTreeView in Build Information with Adwaita widgets
I originally put the information in the `AdwActionRow`'s subtitle, but then the important part (in the subtitle) becomes a dim label, which seems wrong.
I also tried an `AdwEntryRow` instead, which makes the title dim instead, but then you get a weird semi-editable appearance, such as the pencil icon, even when the field is not editable.

So this appearance with a label on the right is similar to the Information section of the General tab in the GTK Inspector.

I also re-arranged the order slightly; it seems like the Meson arguments are the least important so I put them last.

Testing Done:
Compiled and opened About dialog.

Reviewed at https://reviews.imfreedom.org/r/2193/