Replace Developer/Translator GtkTreeView by Adwaita in About dialog
This is just a somewhat larger list than before, but that seems fine.

Theoretically, we could put in any widget here, like a `PidginContact`-esque row with an avatar, etc., but we don't have that information. But it's a possibility for developers, etc., though might be a bit overkill.

Testing Done:
Opened About dialog and checked that Developer/Translator tabs had something in them.

Reviewed at https://reviews.imfreedom.org/r/2194/