Implement SASL for IRCv3 using GSASL
Right now we really only tested PLAIN but SCRAM _might_ work as well as
external. But we'll need to connect to an external server to test that stuff as
it's really a bit of work to get set up locally.

Testing Done:
Connected to a local ergo instance and verified everything there using the PLAIN mechanism.

Reviewed at https://reviews.imfreedom.org/r/2185/