Remove the account lookup on account new
This is used to stop people from creating duplicate accounts. But this is bad because it hides the fact that they user recreated an existing account and will throw out new options in the UI and stuff like that.

This just disables the check for now, and I have a few more review requests I'm putting together to make this a better user experience.

Testing Done:
Ran the unit tests and created a duplicate account.

Reviewed at https://reviews.imfreedom.org/r/2214/