Make the DisplayWindow use a GtkListView
The traditional pattern of using a TreeView and Notebook/stack has been
removed. This now use a ListView with a model that contains the child widgets,
and an AdwBin where we swap the children in and out.

I didn't implement everything, as now that this window is a bit more mixed
content, things aren't as clear cut. I left some TODO's in the code describing
these scenarios.

Testing Done:
Tested opening and closing chats in the ui. Verified the keybinds work for navigating everything thing.

Bugs closed: PIDGIN-17752

Reviewed at https://reviews.imfreedom.org/r/2227/