Add a name-for-display property to PurplePerson
This will first return the alias of the person if set. If there is no alias set
on the person but there is a priority contact info, the value of
purple_contact_info_get_name_for_display will be called on the priority contact
info, otherwise NULL will be returned.

Testing Done:
Verified everything looked right in the contact list and verified everything looked alright in the docs and ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2205/