XMPP: Remove STUN lookups
While this worked in some regard, it was only refreshed when an account
connected which will lead to caching issues over time.

Also as we're trying to migrate all of the NAT traversal stuff to traversity,
this code was a blocker for removing the STUN code from libpurple itself.

Testing Done:
Connected an XMPP account without issue.

Reviewed at https://reviews.imfreedom.org/r/2173/