Add purple_strmatches and move purple_person_matches and purple_contact_info_matches to it
This uses a naive algorithm to do fuzzy matching but appears to be good enough
for now. We can look at something more complicated later if the performance of
this is bad enough.

Testing Done:
Ran the unit tests and tested the search in the contact list.

Bugs closed: PIDGIN-17737

Reviewed at https://reviews.imfreedom.org/r/2212/