Remove PidginBlist
PidginContactList is functional enough for us to start adding features and
since GtkTreeView is being deprecated, now seemed like a great time to purge
PidginBlist.

There's a few remaining pieces in gtkblist.[ch] relating to the join chat and
im buddy dialogs, as well as the PurpleBlist subclass we need to work out yet.

I also removed the gtkbuddynote plugin as all it did was add the buddy note to
the tooltip. My future plans here will have this built in to the ui because a
few protocols support this already and we can mock it for those that don't.

Testing Done:
Launched Pidgin 3 and verified stuff worked!

Bugs closed: PIDGIN-17751

Reviewed at https://reviews.imfreedom.org/r/2228/