Avoid some GWarnings when selecting away from contacts
When we move the DisplayWindow to GtkListView, we moved to using an AdwBin as
the container widget and just set the child to what should be displayed. When
this happens to a ListView, each item gets deleted and for some reason the
bound property closures in the factory get called with an item that is set to
NULL. So if we get passed NULL, we just need to bail.

Testing Done:
Ran under `gdb` with `G_DEBUG=fatal-warnings` and verified that it didn't break when I switched the display window from contacts to conversations.

Reviewed at https://reviews.imfreedom.org/r/2233/