Remove a bunch stuff from PidginConversation
This removes the memberlist, the typing stuff, the send history, as well as
the history loading. The memberlist will be it's on widget in the near
future with libpurple managing the list model, send history will talk to the
history api, and the rest either is or will be implemented in talkatu.

Testing Done:
Joined a MUC and verified that the chat functioned as expected.

Reviewed at https://reviews.imfreedom.org/r/2237/