Display the preferences window as a singleton from the application
This is how we handle displaying other dialogs and apparently the preferences
window just got missed. We also make it transient for whatever window is active
which means it won't fill an entire cell in a tiling window manager.

Testing Done:
Opened the preferences dialog in i3 and verified it was no longer a full tile and that everything worked as expected.

Reviewed at https://reviews.imfreedom.org/r/2242/