disco: Move chat name to Description column
I find it weird that the name of chats ends up in the Name column, so that you don't get the JID.

So move it to the Description, and put the JID back in the Name column.

Testing Done:
Browsed to `pidgin.im`, and expanded `conference.pidgin.im`.

Reviewed at https://reviews.imfreedom.org/r/2275/