Convert XMPP Service Discovery to GtkColumnView
As with the other ports to GtkColumnView, this drops the popover and the tooltip.

Connecting to the expanded property (in order to lazy-load children) is a bit weird, but seems to be the only way I could find to do it.

Testing Done:
Browsed `pidgin.im`, double-clicked `conference.pidgin.im`/clicked the expander, and the rooms were loaded. Double-clicked all the rooms and confirmed the Add Chat dialog was correctly filled. Also clicked the Add button after selecting one and confirmed the same.

Bugs closed: PIDGIN-17747

Reviewed at https://reviews.imfreedom.org/r/2256/