Fix property warning when requesting an int field
The `GtkSpinButton` does not have an `activates-default` property, and so this
warned when creating an integer request field.

But `setup_entry_field` is only called in two places, so we can inline the
property setting to the place that needs it.

Testing Done:
Opened a request with an int field.

Reviewed at https://reviews.imfreedom.org/r/2291/