Close wait request when dialog is destroyed
Because the request was previously only closed when the Cancel button was
clicked, the request was dangling if the user clicked the window Close button,
or pressed Escape.

Testing Done:
Opened a wait request, then closed the dialog with Escape/the close button, and confirmed the cancel callback was triggered.

Reviewed at https://reviews.imfreedom.org/r/2293/