Port list request fields from GtkTreeView to GtkListView
This uses the same "trick" used elsewhere of a plain `GObject` to store the
data we need in the list model.

Note this does not change datasheets.

Testing Done:
Opened a request with a list field and confirmed that the selected options were correct upon accepting the dialog.

Bugs closed: PIDGIN-17750

Reviewed at https://reviews.imfreedom.org/r/2294/