Add demo of requests API to demo protocol
This provides actions that trigger all of the request API, except for datasheets, which are quite complex.

We could do something fancy like have this trigger something with the demo contacts, but all this does is print a debug message with the results.

Testing Done:
Opened all added menu items and confirmed that the correct thing was printed out as input, or that a cancellation was printed.

Reviewed at https://reviews.imfreedom.org/r/2289/