Convert PurpleRequestField into a GObject
This is just a straight conversion with not much attempt to make things nice.

Also, this leaves `PurpleRequestField` as a multi-type object, but derivable, so we can change to a bunch of subclasses after.

Testing Done:
Compiled and opened Request Fields from Demo protocol.

Reviewed at https://reviews.imfreedom.org/r/2320/