Make PurpleRequestFieldImage into a GObject
I'm not sure taking a `char*` and `gsize` is the best constructor; we can probably move this to `GdkPixbuf` or something now. Plus the properties would be nicer, and then it could change images on the fly. But that can happen later.

Testing Done:
Compiled and opened Request Fields from the Demo protocol

Reviewed at https://reviews.imfreedom.org/r/2339/