Make required, but unfilled, fields invalid
And we no longer assume that unrequired-and-unfilled fields are valid. Since there's no default validation for strings (which are currently the only fields that can be unfilled), this is automatically allowed anyway.

If still needed, this can be left up to the subclass-specific or user-supplied validation instead.

Testing Done:
Compiled and ran `ninja test`.

Reviewed at https://reviews.imfreedom.org/r/2351/