Fix new request warning in New IM dialog
Setting the validator triggers the field's `valid` property notification, which triggers the group to refresh its validity, which fetches the `valid` property and runs the validator.

But the validator depends on the account field which hasn't been created yet, which causes a type-check warning.

So set the validator after all fields are made, and also change the data argument so the dependency is clearer.

Testing Done:
Compiled, and opened the New IM dialog.

Reviewed at https://reviews.imfreedom.org/r/2358/