Remove all of the current Mood API
The functionality of the mood api is being moved into PurplePresence via the
soon to be added Emoji and Message properties. As such we have no need for the
existing API which is why we are removing all of it.

Testing Done:
Ran `ninja pidgin-pot all tests` successfully and verified there were no new warnings.

Reviewed at https://reviews.imfreedom.org/r/2366/