Wrap Request Fields dialog in Adwaita widgets
This is the straightforward conversion of simply putting the existing widgets in corresponding `AdwActionRow`. All the different field types are not yet changed.

As an example, I have converted the boolean field to a `GtkSwitch`.

Testing Done:
Compiled and opened Request Fields from the Demo protocol.

Reviewed at https://reviews.imfreedom.org/r/2355/