Add "new-status" properties to PurplePresence with fallbacks were applicable
This adds the new `emoji` and `mobile` properties to `PurplePresence`. It also makes `message` read/write and it fallsback to the active status's message if the internal message is false. Additionally the primitive property was moved `PurpleStatusPrimitive` to `PurplePresencePrimitive`.

Testing Done:
Ran the updated unit tests and connected a demo account and verified that the expected status messages were displayed.

Reviewed at https://reviews.imfreedom.org/r/2372/