Remove the Facebook plugin
This hasn't been used in forever and is behind on patches from the purple2
version so this for sure is not going to work. This will also stop people from
tryingt to update the purple2 version for this code as that's a horrible idea.

Testing Done:
had the turtles do their thing

Reviewed at https://reviews.imfreedom.org/r/2393/